stage 3 complete form done, new look done
