+++ app/globals.css
@@ -3,13 +3,18 @@
 :root {
 	--background: #105693;
 	--foreground: #d1d3df;
-	--destructive: #7f0808;
+	--destructive: #d41e1e;
 	--primary: #006ee3;
 	--primary-selected: #007aff;
 	--secondary: #ff9500;
 	--surface: #f2f2f7;
 }
-
+@theme inline {
+	--color-background: var(--background);
+	--color-foreground: var(--foreground);
+	--color-destructive: var(--destructive);
+	
+}
 @keyframes fadeBorderIn {
 	0% {
 		border-bottom-color: transparent;
