+++ components/ui/containers/content-container-dark.tsx
@@ -4,7 +4,7 @@ export default function ContentContainerDark({
 	children: React.ReactNode
 }) {
 	return (
-		<div className="w-full max-w-4xl rounded-xl border border-slate-800 bg-slate-900 p-8 text-center shadow-xl">
+		<div className="w-full max-w-4xl rounded-xl border border-slate-800 bg-slate-900 p-4 lg:p-6 text-center shadow-xl">
 			{children}
 		</div>
 	)
